+++ apps/astarte_appengine_api/CHANGELOG.md
@@ -16,11 +16,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change logs format to logfmt.
 
 ## [0.10.2] - Unreleased
+### Added
+- Add timestamp field to channel events.
+
 ### Fixed
 - Fix invalid dates handling, they should not cause an internal server error.
 - Gracefully handle existing aliases instead of returning an internal server error.
 - Fix querying object aggregated interface with explicit timestamp, use value_timestamp to avoid
 an internal server error.
+- Device details now show false in the connected field for never connected devices (null was
+  returned before)
 
 ## [0.10.1] - 2019-10-02
 ### Fixed